Keep SplitPanel beside Tab, not inside the pane tree
Tabs/Panes/ is the pane tree model with no XAML in it; SplitPanel is
the Panel that renders that tree and is owned by Tab, so it belongs
next to Tab.h.
