+++ App/Tabs/SplitPanel.idl
